requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-filter==23.2
django-quill-editor==0.1.40
djangorestframework==3.14.0
drf-nested-routers==0.93.4
Markdown==3.4.3
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
pillow=9.5.0
